Unify drop implementations.
